Move new files for ExternalInitiator into core/
