+++ core/web/external_initiators_controller_test.go
@@ -3,8 +3,8 @@ package web_test
 import (
 	"testing"
 
+	"github.com/smartcontractkit/chainlink/core/store/models"
 	"github.com/smartcontractkit/chainlink/internal/cltest"
-	"github.com/smartcontractkit/chainlink/store/models"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
